+++ src/containerapp/azext_containerapp/_clients.py
@@ -663,7 +663,7 @@ class ConnectedEnvCertificateClient():
         return r.json()
 
     @classmethod
-    def delete_certificate(cls, cmd, resource_group_name, name, certificate_name):
+    def delete_certificate(cls, cmd, resource_group_name, name, certificate_name, no_wait=False):
         management_hostname = cmd.cli_ctx.cloud.endpoints.resource_manager
         sub_id = get_subscription_id(cmd.cli_ctx)
         url_fmt = "{}/subscriptions/{}/resourceGroups/{}/providers/Microsoft.App/connectedEnvironments/{}/certificates/{}?api-version={}"
@@ -675,7 +675,16 @@ class ConnectedEnvCertificateClient():
             certificate_name,
             cls.api_version)
 
-        return send_raw_request(cmd.cli_ctx, "DELETE", request_url, body=None)
+        r = send_raw_request(cmd.cli_ctx, "DELETE", request_url, body=None)
+
+        if no_wait:
+            return  # API doesn't return JSON (it returns no content)
+        elif r.status_code == 202:
+            operation_url = r.headers.get(HEADER_LOCATION)
+            poll_results(cmd, operation_url)
+            logger.warning('Certificate %s was successfully deleted', certificate_name)
+
+        return
 
     @classmethod
     def check_name_availability(cls, cmd, resource_group_name, name, name_availability_request):
@@ -723,7 +732,7 @@ class ConnectedEnvDaprComponentClient():
         return r.json()
 
     @classmethod
-    def delete(cls, cmd, resource_group_name, environment_name, name):
+    def delete(cls, cmd, resource_group_name, environment_name, name, no_wait=False):
         management_hostname = cmd.cli_ctx.cloud.endpoints.resource_manager
         sub_id = get_subscription_id(cmd.cli_ctx)
         url_fmt = "{}/subscriptions/{}/resourceGroups/{}/providers/Microsoft.App/connectedEnvironments/{}/daprComponents/{}?api-version={}"
@@ -735,7 +744,15 @@ class ConnectedEnvDaprComponentClient():
             name,
             cls.api_version)
 
-        send_raw_request(cmd.cli_ctx, "DELETE", request_url)
+        r = send_raw_request(cmd.cli_ctx, "DELETE", request_url)
+
+        if no_wait:
+            return  # API doesn't return JSON (it returns no content)
+        elif r.status_code == 202:
+            operation_url = r.headers.get(HEADER_LOCATION)
+            poll_results(cmd, operation_url)
+            logger.warning('Dapr component %s was successfully deleted', name)
+
         return
 
     @classmethod
@@ -788,7 +805,7 @@ class ConnectedEnvStorageClient():
     api_version = PREVIEW_API_VERSION
 
     @classmethod
-    def create_or_update(cls, cmd, resource_group_name, env_name, name, storage_envelope):
+    def create_or_update(cls, cmd, resource_group_name, env_name, name, storage_envelope, no_wait=False):
         management_hostname = cmd.cli_ctx.cloud.endpoints.resource_manager
         sub_id = get_subscription_id(cmd.cli_ctx)
         url_fmt = "{}/subscriptions/{}/resourceGroups/{}/providers/Microsoft.App/connectedEnvironments/{}/storages/{}?api-version={}"
@@ -802,10 +819,17 @@ class ConnectedEnvStorageClient():
 
         r = send_raw_request(cmd.cli_ctx, "PUT", request_url, body=json.dumps(storage_envelope))
 
+        if no_wait:
+            return r.json()
+        elif r.status_code == 201:
+            operation_url = r.headers.get(HEADER_AZURE_ASYNC_OPERATION)
+            poll_status(cmd, operation_url)
+            r = send_raw_request(cmd.cli_ctx, "GET", request_url)
+
         return r.json()
 
     @classmethod
-    def delete(cls, cmd, resource_group_name, env_name, name):
+    def delete(cls, cmd, resource_group_name, env_name, name, no_wait=False):
         management_hostname = cmd.cli_ctx.cloud.endpoints.resource_manager
         sub_id = get_subscription_id(cmd.cli_ctx)
         url_fmt = "{}/subscriptions/{}/resourceGroups/{}/providers/Microsoft.App/connectedEnvironments/{}/storages/{}?api-version={}"
@@ -817,7 +841,14 @@ class ConnectedEnvStorageClient():
             name,
             cls.api_version)
 
-        send_raw_request(cmd.cli_ctx, "DELETE", request_url)
+        r = send_raw_request(cmd.cli_ctx, "DELETE", request_url)
+
+        if no_wait:
+            return  # API doesn't return JSON (it returns no content)
+        elif r.status_code == 202:
+            operation_url = r.headers.get(HEADER_LOCATION)
+            poll_results(cmd, operation_url)
+            logger.warning('Storage %s was successfully deleted', name)
 
         return