+++ src/mcc/azext_mcc/aaz/latest/mcc/isp/resource/_list.py
@@ -776,9 +776,7 @@ class _ListHelper:
         )
 
         details = _schema_error_detail_read.details
-        details.Element = AAZObjectType(
-            flags={"read_only": True},
-        )
+        details.Element = AAZObjectType()
         cls._build_schema_error_detail_read(details.Element)
 
         _schema.additional_info = cls._schema_error_detail_read.additional_info