+++ src/maintenance/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+1.7.0b2
+++++++
+* Added list of allowed classifications in description for Maintenance Configuration
+
 1.7.0b1
 ++++++
 * Changed incorrect parameter name